+++ src/lufah/commands/core/create_group.py
@@ -7,11 +7,11 @@ async def do_create_group(args: argparse.Namespace):
     "Create group if it does not exist."
     client = args.client
     await client.connect()
-    await client.create_group(client.group)
+    await client.create_group(args.group)
 
 
 async def do_delete_group(args: argparse.Namespace):
     'Delete group if it exists, is not "", is paused, and has no units.'
     client = args.client
     await client.connect()
-    await client.delete_group(client.group)
+    await client.delete_group(args.group)